♻️(identite-domain) move identite_errors to domain
